update tests. add install config with commands
