+++ .gitignore
@@ -1,3 +1,6 @@
 __pycache__
 venv
 *.db
+ghistintime.egg-info/dependency_links.txt ghistintime.egg-info/entry_points.txt ghistintime.egg-info/PKG-INFO ghistintime.egg-info/SOURCES.txt ghistintime.egg-info/top_level.txt
+ghistintime.egg-info/dependency_links.txt ghistintime.egg-info/entry_points.txt ghistintime.egg-info/PKG-INFO ghistintime.egg-info/SOURCES.txt ghistintime.egg-info/top_level.txt
+ghistintime.egg-info
