+++ README.md
@@ -1,15 +1,18 @@
 # ghistintime
 
 ```bash
+# install
+pip install git+ssh://git@github.com/possemeeg/ghistintime.git@dev-upgrade
+
 # For bashrc.sh or bash_profile.sh
-export PROMPT_COMMAND='python3.7 /path/to/ghist.py --database $HOME/.ghist.db put "$(fc -ln -0)"'
+export PROMPT_COMMAND=gh_put "$(fc -ln -0)"
 
-export GHIST_SCRIPT="$HOME/Development/scripts/ghist.py"
-export GHIST_DB="$HOME/.ghist.db"
-
-alias gh="python3.7 $GHIST_SCRIPT --database $GHIST_DB get"
-alias gha="python3.7 $GHIST_SCRIPT --database $GHIST_DB ass"
-alias ghr="python3.7 $GHIST_SCRIPT --database $GHIST_DB ref"
-alias ghe="python3.7 $GHIST_SCRIPT --database $GHIST_DB ex"
 ```
+
+## Config
+- Config is stored in ~/.config/ghistintime/config.cfg
+- Currently this is only the database.
+- Database defaults to ~/.config/ghistintime/ghist.db
+- Thus if no config is created, all that you'll get is `~/.config/ghistintime/ghist.db`.
+- Thus (II) make sure your existing db is copied to `~/.config/ghistintime/ghist.db`.
